Port over the old Health webapp (#2)
I copy over the old javascript code from my original Health app into the client-javascript directory, then rebuild that into Typescript in the client directory. This is a rough pass at the typescript and a lot more work is still needed to get that filled out completely.

As of this commit, the code cannot work together as there are no credentials and there are endpoints missing. Both of these need to be filled out on the server side before this app can really get moving.

* Copy the client application over from my old health app
* Move the server into a dedicated directory
* Add node9 to the environment
* Create a Typescript version of the existing application
* Start building with webpack
* Add an index.html to anchor the webapp
* Update paths for the server CI build