Add a bunch of options to get the build right
Lots of webpack configuration that I worked out for my article on
setting up a typescript project. Then I added better error handling for
StaticFile so that I could tell what file path was failing. Possibly
need an improved error class for file handling errors which includes the
path to the file that is in trouble.