Fix a bug with timezone handling in DateTimeTz
Because it really is hard to get right...

"2019-05-15T15:30:00Z America/New_York" was being displayed as
"2019-05-15T11:30:00Z America/New_York". All of the correct information
was present, but the formatting function was displaying the time in the
local time zone, which is incorrect. It should display the time in Z,
especially with Z present in the resulting string.