Set up the Ubuntu build (#55)
Switch from CircleCI build to old-fashioned make with dockers.

In order to get a container that has Rust, Node, and FPM, I end up with 3GB container (1GB compressed). CircleCI was starting to have troubles with that. Maybe I'll return to Circle eventually, but right now that is very costly.