Set up a date_selector component
This is a lot of work to get the component set up, to ensure that the
calendar only gets displayed when desired, and then to figure out how to
make the selector independent of the context, but still route messages
through to the context.

I hope I'm not making a disaster.